Update list of selected tests to run.
Modified RideTestSelectedForRunningChanged message to hold the info re a single test being selected/unselected. (before it was always passing the full list of tests).

TestCaseFileController to propagate RideTestSelectedForRunningChanged to notify about test selection. TestCaseNodeHandler and TestRunnerPlugin to listen to these events. TestRunnerPlugin to keep the list of selected tests, updated.

TestCaseFileController to iterate test cases using iter(). So that TestCaseController(s) are created on the fly. They need to be propagated in the event.

Inhibit TestSelectionController to send the same events. That class has no utility as of now.